xdg-shell: Introduce xdg_positioner
xdg_positioner is a method for declarative positioning of child
surfaces, such as xdg_popup and xdg_tooltip surfaces. A client creates
a description of a positioning logic using the xdg_positioner
interface. The xdg_positioner object is then used when creating a
xdg_popup or xdg_tooltip for describing how the child surface should be
positioned in relation to the parent surface.

Signed-off-by: Jonas Ådahl <jadahl@gmail.com>